fix general app flow, fix adding new entries
