+++ app/src/main/res/layout/activity_nextexpectedcycle.xml
@@ -7,20 +7,20 @@
     tools:context=".activities.NewCycleActivity">
 
     <Button
-        android:id="@+id/abortbtn"
+        android:id="@+id/abortButton1"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:text="@string/button_abort"
         android:textSize="18sp"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/addRangeButton"
+        app:layout_constraintEnd_toStartOf="@+id/addExpectedButton"
         app:layout_constraintHorizontal_weight="1"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.976" />
 
     <Button
-        android:id="@+id/addRangeButton"
+        android:id="@+id/addExpectedButton"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:text="@string/button_finish"
@@ -28,7 +28,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_weight="1"
-        app:layout_constraintStart_toEndOf="@+id/abortbtn"
+        app:layout_constraintStart_toEndOf="@+id/abortButton1"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.976" />
