+++ app/src/main/res/values/strings.xml
@@ -1,8 +1,8 @@
 <resources>
     <string name="app_name">MensTracker</string>
     <string name="button_abort">Abort</string>
-    <string name="button_next">Next</string>
-    <string name="button_finish">Finish</string>
+    <string name="button_next">Add new</string>
+    <string name="button_finish">Add expected</string>
     <string name="button_addnewperiod">Add new period</string>
     <string name="expectation_list">Expectation: %1$s</string>
     <string name="reality_list">Reality: %1$s</string>
